Update packages and Add test

+++ packages/protocol/contracts/vesting/vesting.sol
@@ -5,6 +5,7 @@ import '@openzeppelin/contracts/utils/Address.sol';
 import '@openzeppelin/contracts/utils/Context.sol';
 import '@openzeppelin/contracts/utils/math/Math.sol';
 import '@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol';
+import 'hardhat/console.sol';
 
 /**
  * @title VestingWallet
@@ -25,7 +26,7 @@ contract VestingWallet is Context {
         uint64 durationSeconds;
         uint64 cliffSeconds;
         uint64 periodSeconds;
-        uint256 lastRelease;
+        uint64 lastRelease;
         uint256 released;
     }
 
@@ -101,10 +102,11 @@ contract VestingWallet is Context {
             block.timestamp >= vesting[_beneficiaryAddress].lastRelease + vesting[_beneficiaryAddress].periodSeconds,
             'EQUITY DAO: Next vesting period not reached'
         );
+
         uint256 releasable = vestedAmount(_beneficiaryAddress, uint64(block.timestamp)) - released(_beneficiaryAddress);
         require(releasable > 0, 'EQUITY DAO: No vested token available');
         vesting[_beneficiaryAddress].released = vesting[_beneficiaryAddress].released + releasable;
-        vesting[_beneficiaryAddress].lastRelease = block.timestamp;
+        vesting[_beneficiaryAddress].lastRelease = uint64(block.timestamp);
         emit ERC20Released(releasable);
         SafeERC20.safeTransfer(IERC20(_token), _beneficiaryAddress, releasable);
     }
@@ -113,14 +115,23 @@ contract VestingWallet is Context {
      * @dev Calculates the amount of tokens that has already vested. Default implementation is a linear vesting curve.
      */
     function vestedAmount(address _beneficiaryAddress, uint64 timestamp) public view virtual returns (uint256) {
-        return _vestingSchedule(_beneficiaryAddress, IERC20(_token).balanceOf(address(this)) + released(_beneficiaryAddress), timestamp);
+        return
+            _vestingSchedule(
+                _beneficiaryAddress,
+                IERC20(_token).balanceOf(address(this)) + released(_beneficiaryAddress),
+                timestamp
+            );
     }
 
     /**
      * @dev Virtual implementation of the vesting formula. This returns the amout vested, as a function of time, for
      * an asset given its total historical allocation.
      */
-    function _vestingSchedule(address _beneficiaryAddress, uint256 totalAllocation, uint64 timestamp) internal view virtual returns (uint256) {
+    function _vestingSchedule(
+        address _beneficiaryAddress,
+        uint256 totalAllocation,
+        uint64 timestamp
+    ) internal view virtual returns (uint256) {
         if (timestamp < start(_beneficiaryAddress)) {
             return 0;
         } else if (timestamp > (start(_beneficiaryAddress) + duration(_beneficiaryAddress))) {
